Implement the `contains` matcher
This matcher has been implemented for all types in the standard library
that have a method called `contains` on them, with the exception of
`EnumSet` which is not stable. Unfortunately, there is no trait that
encompases all collections, so the trait had to be implemented manually.
We could potentially implement this via `Iterator` instead, but that
wouldn't necessarily match the semantics of each collection.

`pattern` is also unstable, which is unfortunate since we're not relying
on it, only the signature of `str#contains`. The "real" impl for this
has been hidden behind a feature, and it simply implements the matcher
for two strings on stable.